requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.3
aiohttp-client-cache==0.11.1
aiosignal==1.3.1
aiosqlite==0.20.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bloonspy==0.9.1
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
itsdangerous==2.2.0
multidict==6.1.0
pillow==10.4.0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.2.3
yarl==1.13.0
